+++ app/src/view/h_recycling.cpp
@@ -90,6 +90,8 @@ HRecyclerFrame::~HRecyclerFrame() {
 void HRecyclerFrame::setDataSource(RecyclingGridDataSource* source) {
     if (this->dataSource) delete this->dataSource;
 
+     // 允许自动加载下一页
+    this->requestNextPage = false;
     this->dataSource = source;
     if (layouted) reloadData();
 }