+++ app/src/view/recycling_grid.cpp
@@ -266,6 +266,8 @@ void RecyclingGrid::addCellAt(size_t index, bool downSide) {
 void RecyclingGrid::setDataSource(RecyclingGridDataSource* source) {
     if (this->dataSource) delete this->dataSource;
 
+    // 允许自动加载下一页
+    this->requestNextPage = false;
     this->dataSource = source;
     if (layouted) reloadData();
 }