+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * add close button for series/collection
 * add lock button for player view
+* add close button for player view
 
 ### Fixed